Create imgproc module for cv processing
